Moved SNS canister.rs tests to their own file.
